+++ src/gui/CustomLNFs.h
@@ -97,3 +97,29 @@ public:
 private:
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (ComboBoxLNF)
 };
+
+class TuningMenuLNF : public ComboBoxLNF
+{
+public:
+    TuningMenuLNF() = default;
+    ~TuningMenuLNF() override = default;
+
+    void drawComboBox (Graphics& g, int width, int height, bool, int, int, int, int, ComboBox& box) override
+    {
+        auto cornerSize = 5.0f;
+        Rectangle<int> boxBounds (0, 0, width, height);
+
+        g.setColour (box.findColour (ComboBox::backgroundColourId));
+        g.fillRoundedRectangle (boxBounds.toFloat(), cornerSize);
+    }
+
+    void positionComboBoxText (ComboBox& box, Label& label) override
+    {
+        auto b = box.getBounds();
+        label.setBounds (b);
+        label.setFont (getComboBoxFont (box).boldened());
+    }
+
+private:
+    JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (TuningMenuLNF)
+};