+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED
+- Added support for custom tunings.
+
 ## [1.0.0] 2021-07-08
 - Initial release.